wip: Test local conda build
So far the recipe builds locally, but fails with QA tests.
The observed issue arises during import. Smth like "could not find
gr::sync_block". The same issue arose with a purely local build when the
system picked up an older pybind11 version. This seems to be not the
case.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>